chore(shared): relocate formatWithIndefiniteArticle to string utils
Move formatWithIndefiniteArticle from src/shared/src/utils/object.js into src/shared/src/utils/string.js so the text-formatting helper lives alongside the rest of the string utilities. Update dependent imports and tests to reflect the new module location.
